Fixes unattached popovers. Closes #7
Fixes an issue where popovers were not attached to the seats they were
supposed to be when the scrollview was zoomed. This was because we were
calculating the rect in the seating chart coordinates system, but were
presenting in a view other than the seating chart. This commit converts
the rectangle to the correct coordinate system before presenting.

Note that we chose to convert the coordinates of the rect rather than
just presenting the popover in the seating chart view. We chose this
method because we weren't sure if presenting a popover in the
contentView of a scrollview would make the popover think that it could
go offscreen, which would be a terrible experience. We would rather just
convert the coordinates to a rect that is entirely visible