Adds deselection of group. Fixes #20
Changes the edit student view to show groups as a section instead of a
popover. This section allows the deselection of groups. Updates the save
method to deal with this new use case.

Also updates our usage of QDateTimeInlineElement to include the new mode
parameter which was added when we updated QuickDialog.