Group section will not show up with no groups.
If there are no groups in a section, then the
TAStudentEditViewController will not longer show an empty section. The
section will only get added if there is at least 1 group. This commit
also rearranges some of the code to fit in a more logical spot, and
removes some comments I accidentaly committed.