Fixes crash with groups and fixes #16
Fixes a major crash with groups that happened when the user didnt press
any buttons in the groups section while they didnt have a group
assigned.

Also fixes an issue (#16) where adding a student changed the index of
the student that had a detailed cell. We solved this by saving the
detailed student before we updated the list, and then setting the
detailedStudentIndex to be the index of the same student as before.